Take into account faults and current aircraft state for wing anti ice. Also prep for engine anti ice.
