+++ fbw-a32nx/src/wasm/systems/a320_systems_wasm/src/lib.rs
@@ -61,6 +61,7 @@ async fn systems(mut gauge: msfs::Gauge) -> Result<(), Box<dyn Error>> {
         (ElectricalBusType::DirectCurrentGndFltService, 15),
     ])?
     .with_auxiliary_power_unit(Variable::named("OVHD_APU_START_PB_IS_AVAILABLE"), 8, 7)?
+    .with_wing_anti_ice()?
     .with_failures(vec![
         (
             21_000,
@@ -277,7 +278,6 @@ async fn systems(mut gauge: msfs::Gauge) -> Result<(), Box<dyn Error>> {
         (34_000, FailureType::RadioAltimeter(1)),
         (34_001, FailureType::RadioAltimeter(2)),
     ])
-    .with_wing_anti_ice()?
     .provides_aircraft_variable("ACCELERATION BODY X", "feet per second squared", 0)?
     .provides_aircraft_variable("ACCELERATION BODY Y", "feet per second squared", 0)?
     .provides_aircraft_variable("ACCELERATION BODY Z", "feet per second squared", 0)?
